MOW-1106: setting vertical alignment of table cells from the quickstylebar

See merge request ggb/geogebra!1646

+++ web/src/main/java/org/geogebra/web/full/gui/util/PopupMenuButtonW.java
@@ -452,8 +452,7 @@ public class PopupMenuButtonW extends MyCJButton
 	 *            {@code Integer}
 	 */
 	public void setSelectedIndex(Integer selectedIndex) {
-		myTable.setSelectedIndex(
-				selectedIndex == null ? -1 : selectedIndex.intValue());
+		myTable.setSelectedIndex(selectedIndex == null ? -1 : selectedIndex);
 		updateGUI();
 	}
